fix tables

+++ rest/src/main/java/ca/deos/store/dao/impl/OrderDaoImpl.java
@@ -2,6 +2,8 @@ package ca.deos.store.dao.impl;
 
 import ca.deos.store.dao.OrderDao;
 import ca.deos.store.entity.Order;
+import org.apache.juli.logging.Log;
+import org.apache.juli.logging.LogFactory;
 import org.hibernate.Session;
 import org.hibernate.criterion.Restrictions;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +19,9 @@ import java.util.List;
 @Repository
 public class OrderDaoImpl implements OrderDao {
 
+    private Log log = LogFactory.getLog(RestaurantDaoImpl.class);
+
+
     @Autowired
     EntityManager em;
 
@@ -29,7 +34,17 @@ public class OrderDaoImpl implements OrderDao {
     }
 
     @Override
-    @Transactional
+    @Transactional(readOnly = true)
+    public List<Order> gesRestaurantOrders(int resId){
+        Session session = em.unwrap(Session.class);
+
+        return session.createCriteria(Order.class)
+                .add(Restrictions.eq("res_id", resId))
+                .list();
+    }
+
+    @Override
+    @Transactional(readOnly = true)
     public Order getOrder(int orderId){
         Session session = em.unwrap(Session.class);
 
@@ -37,11 +52,13 @@ public class OrderDaoImpl implements OrderDao {
     }
 
     @Override
-    @Transactional
+    @Transactional(readOnly = true)
     public List<Order> getUserOrders(String userId) {
         Session session = em.unwrap(Session.class);
 
-        return session.createCriteria(Order.class).add(Restrictions.eq(userId, userId)).list();
+        return session.createCriteria(Order.class)
+                .add(Restrictions.eq("user_id", userId))
+                .list();
 
     }
 
@@ -50,7 +67,14 @@ public class OrderDaoImpl implements OrderDao {
     public void saveOrUpdateOrder(Order order){
         Session session = em.unwrap(Session.class);
 
-        session.saveOrUpdate(order);
+        if(order.getId() > 0){
+            session.update(order);
+            log.info("Updating existing order");
+        } else {
+            session.save(order);
+            log.info("Updating existing order");
+        }
+        log.info("Order saved with id: " + order.getId());
     }
 
     @Override
@@ -59,7 +83,8 @@ public class OrderDaoImpl implements OrderDao {
         Session session = em.unwrap(Session.class);
 
         Order order = (Order) session.load(Order.class, orderId);
-        session.delete(order);
-
+        if(order != null) {
+            session.delete(order);
+        }
     }
 }
